backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
starlette
pydantic>=2.10.0,<3.0
pydantic-settings==2.5.2
typing-extensions

azure-functions==1.20.0
azurefunctions-extensions-http-fastapi==1.0.1
httpx==0.27.2
requests
msal
pyjwt[crypto]

openai
tiktoken
langchain
langgraph
llama-index

azure-identity==1.17.1
azure-keyvault-secrets==4.9.0
azure-search-documents==11.6.0b6
azure-mgmt-resource
azure-mgmt-network
azure-storage-blob[aio]==12.23.1
azure-data-tables==12.6.0
azure-servicebus==7.12.2
azure-cosmos==4.3.0
croniter==1.3.8

pyodbc==5.3.0
dsnparse
